// src/types.d.ts
type Product = {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
};

type ProductsData = {
  page: number;
  limit: number;
  totalPages: number;
  previousPage: boolean;
  nextPage: boolean;
  totalItems: number;
  currentPageItems: number;
  data: Product[];
};

type ApiResponse = {
  statusCode: number;
  data: ProductsData;
  message: string;
  success: boolean;
};

export { ApiResponse, Product, ProductsData };